Refactor : enhance the build maven report shell (#680)

+++ .github/scripts/build-maven-report.sh
@@ -32,13 +32,15 @@ echo "4. process maven stage"
 ./mvnw site:stage
 
 echo "5. upload the maven report"
-rm -rf ${SHARDINGSPHERE_DOC_PATH}/statistics/staging
-mv target/staging  ${SHARDINGSPHERE_DOC_PATH}/statistics/
-git config --global user.email "dev@shardingsphere.apache.org"
-git config --global user.name "shardingsphere"
-cd ${SHARDINGSPHERE_DOC_PATH}
-git add .
-export TZ="Asia/Shanghai"
-dateStr=`date "+%Y-%m-%d %H:%M:%S %Z"`
-git commit -m  "Update shardingsphere maven report at $dateStr."
-git push
+if [ -d "$directory" ] && [ -n "$(ls -A "$directory")" ]; then
+  rm -rf ${SHARDINGSPHERE_DOC_PATH}/statistics/staging
+  mv target/staging  ${SHARDINGSPHERE_DOC_PATH}/statistics/
+  git config --global user.email "dev@shardingsphere.apache.org"
+  git config --global user.name "shardingsphere"
+  cd ${SHARDINGSPHERE_DOC_PATH}
+  git add .
+  export TZ="Asia/Shanghai"
+  dateStr=`date "+%Y-%m-%d %H:%M:%S %Z"`
+  git commit -m  "Update shardingsphere maven report at $dateStr."
+  git push
+fi
